+++ photon-model/src/main/java/com/vmware/photon/controller/model/constants/ReleaseConstants.java
@@ -23,4 +23,5 @@ public class ReleaseConstants {
     public static final int RELEASE_VERSION_0_6_6 = 66;
     public static final int RELEASE_VERSION_0_6_10 = 610;
     public static final int RELEASE_VERSION_0_6_11 = 611;
+    public static final int RELEASE_VERSION_0_6_12 = 612;
 }